client/windows/requirements.txt
pystray>=0.19.0
Pillow>=9.0.0
requests>=2.28.0
